Move data on contact page to forretag page, Add icon Grid for forateg page
